+++ cuda/unittest/FunctorPoint_tests.cpp
@@ -18,17 +18,17 @@ TEST(FunctorPoint, naive) {
     vector<double> ys{0, 1, 2, 3, 4};
     GeometryVector left_points;
     GeometryVector right_points;
-    ST_Point(xs.data(), ys.data(), xs.size(), left_points);
+    ST_Point(xs.data(), ys.data(), (int)xs.size(), left_points);
     for(auto &x: xs) {
         x = -x;
     }
     for(auto &y: ys) {
         y = -y;
     }
-    ST_Point(xs.data(), ys.data(), xs.size(), right_points);
+    ST_Point(xs.data(), ys.data(), (int)xs.size(), right_points);
     vector<double> distance(xs.size());
     ST_Distance(left_points, right_points, distance.data());
-    for(int i = 0; i < xs.size(); ++i) {
+    for(size_t i = 0; i < xs.size(); ++i) {
         auto std = (xs[i] * xs[i] + ys[i] * ys[i]) * 4;
         auto res = distance[i] * distance[i];
         ASSERT_DOUBLE_EQ(res, std);
